Updated mirror.

+++ core/lib/config/src/configs/eth_client.rs
@@ -7,7 +7,7 @@ use crate::envy_load;
 #[derive(Debug, Deserialize, Clone, PartialEq)]
 pub struct ETHClientConfig {
     /// Numeric identifier of the L1 network (e.g. `9` for localhost).
-    pub chain_id: u8,
+    pub chain_id: u64,
     /// Address of the Ethereum node API.
     pub web3_url: String,
 }
